[Refactor] #34 delete index.ts
[Refactor] #34 delete index.ts
