--- src/components/organisms/index.ts
@@ -1,4 +1,0 @@
-export { default as Header } from './Header';
-export { default as MypageSlide } from './MypageSlide';
-export { default as RoutineCard } from './RoutineCard';
-export { SignupCard } from './SignupCard';
